Data can be sent to sql server
Added tons of logic for sql server. Updated the main (it handles the connection on app open and closed now). I also made the connections very very efficient because the websockets are now handled per client and not per request, and it is always open for each client. Added better handling for this. Finally, added the logic for sending it to the sql server. This is good because now we know how to integrate it with each webpage :) (which I haven't done because yk we need the app in its app form)